[Horus] Fixes #4205 (#4210)
[Horus] Fixes #4205

+++ radio/src/targets/taranis/board.h
@@ -457,10 +457,10 @@ uint8_t eepromIsTransferComplete();
 void debugPutc(const char c);
 
 // Telemetry driver
-void telemetryPortInit(uint32_t baudrate, int mode);
+void telemetryPortInit(uint32_t baudrate, uint8_t mode);
 void telemetryPortSetDirectionOutput(void);
 void sportSendBuffer(uint8_t * buffer, uint32_t count);
-int telemetryGetByte(uint8_t * byte);
+uint8_t telemetryGetByte(uint8_t * byte);
 extern uint32_t telemetryErrors;
 
 // Audio driver
